7-1 edits
I am having trouble getting the individual background colors attached
to each page.  I creating multiple classes with the unique color I
wanted and then assigning that to the individual pages, but because the
bottom class always overrides the top one, that bottom color would
apply to all of the pages.